attempt to fix lab3..

+++ group_9/sidelnikova.sveta/lab3/src/main/java/tasks/SoldDrinkCollectionAnalyzer.java
@@ -12,5 +12,5 @@ public interface SoldDrinkCollectionAnalyzer {
 
     Set<DrinkType> getNotOrderedDrinksLastThreeMonths(@NotNull Collection<SoldDrink> collection);
 
-    int getCappuccinoOrdersCount(@NotNull Collection<SoldDrink> collection);
+    long getCappuccinoOrdersCount(@NotNull Collection<SoldDrink> collection);
 }
